move some code out of scripts directory
